requirements.txt
# API
fastapi==0.115.12
uvicorn==0.42.0
python-dotenv==1.2.2
pydantic==2.12.5

# ML / Embeddings
sentence-transformers==5.3.0
torch==2.11.0

# Vector DB
chromadb==1.5.5

# LLM
langchain-groq
langchain-core

# Data
pandas==3.0.2
numpy==2.4.4
scikit-learn==1.8.0

# Notebooks / EDA
matplotlib
seaborn
plotly
umap-learn
lxml

# Streamlit
streamlit==1.56.0